Move new password page into authentication
